Addition of topoplot in the right hand corner
As a way to continue improving the program, I added the right-click topoplot display at the right corner. Makes for better UX. I wanted to have it on always in the function mouse_motion, but it was crashing when scrolling, and it was a bit slow.
The display couldn't be done for component plots because the eegplot_w2 doesn't receive the full EEG data. If it did, it would have to be remodeled. Couldn't find an easy work around it, since the function is recalled various times inside itself. Sad.