Update eegplot_w2.m
New feature added: Matrix plot! It was hard to do.
Also: removed a bug on line 2265/2252. Used to be a bug fixing code, now it Crashes when plotting sometimes. Something to do with the limits of the plot.
Changed some nomenclature on TBT
Reorganized and cleaned g.winrej.
Fixed merge-trials functions bugs!
Added comments on a lot of functions