Create pop_viewprops3_par.m
Newly modified function: now allows for plotting of the ERPs. Both projected to a channel and for specific epochs, or the normal component ERP.

variable added is fig_opts. fig_opts{1} is 1 (normal headplots), 2 (projected to channel and with specific eventor 3 for classic ERP of the component_props. if choice is 2, it requires a a channel and an event.